requirements.txt
python-telegram-bot==20.*
anthropic
python-dotenv
fastapi
uvicorn